managed_exec: extract spawn/observe/kill into separate modules
